streamlit/requirements.txt
altair==5.5.0 ; python_version >= "3.11"
attrs==25.3.0 ; python_version >= "3.11"
azure-core==1.34.0 ; python_version >= "3.11"
azure-cosmos==4.9.0 ; python_version >= "3.11"
blinker==1.9.0 ; python_version >= "3.11"
cachetools==5.5.2 ; python_version >= "3.11"
certifi==2025.6.15 ; python_version >= "3.11"
charset-normalizer==3.4.2 ; python_version >= "3.11"
click==8.2.1 ; python_version >= "3.11"
colorama==0.4.6 ; python_version >= "3.11" and platform_system == "Windows"
gitdb==4.0.12 ; python_version >= "3.11"
gitpython==3.1.44 ; python_version >= "3.11"
idna==3.10 ; python_version >= "3.11"
jinja2==3.1.6 ; python_version >= "3.11"
jsonschema-specifications==2025.4.1 ; python_version >= "3.11"
jsonschema==4.24.0 ; python_version >= "3.11"
markupsafe==3.0.2 ; python_version >= "3.11"
narwhals==1.43.0 ; python_version >= "3.11"
numpy==2.3.0 ; python_version >= "3.11"
packaging==24.2 ; python_version >= "3.11"
pandas==2.3.0 ; python_version >= "3.11"
pillow==11.2.1 ; python_version >= "3.11"
protobuf==6.31.1 ; python_version >= "3.11"
pyarrow==20.0.0 ; python_version >= "3.11"
pydeck==0.9.1 ; python_version >= "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.11"
pytz==2025.2 ; python_version >= "3.11"
referencing==0.36.2 ; python_version >= "3.11"
requests==2.32.4 ; python_version >= "3.11"
rpds-py==0.25.1 ; python_version >= "3.11"
six==1.17.0 ; python_version >= "3.11"
smmap==5.0.2 ; python_version >= "3.11"
streamlit==1.45.1 ; python_version >= "3.11"
tenacity==9.1.2 ; python_version >= "3.11"
toml==0.10.2 ; python_version >= "3.11"
tornado==6.5.1 ; python_version >= "3.11"
typing-extensions==4.14.0 ; python_version >= "3.11"
tzdata==2025.2 ; python_version >= "3.11"
urllib3==2.4.0 ; python_version >= "3.11"
watchdog==6.0.0 ; python_version >= "3.11" and platform_system != "Darwin"
